requirements.txt
aiohttp==3.6.2
alembic==1.3.0
amqp==2.5.2
apache-airflow==1.10.6
apispec==3.1.0
appdirs==1.4.3
argcomplete==1.10.0
arrow==0.15.4
aspy.yaml==1.3.0
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
black==19.3b0
bs4==0.0.1
cached-property==1.5.1
celery==4.4.0
certifi==2019.9.11
cffi==1.13.2
cfgv==2.0.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.30
cryptography==2.8
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
graphviz==0.13.2
gunicorn==19.9.0
identify==1.4.8
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.23
importlib-resources==1.0.2
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
json-merge-patch==0.2
jsonschema==3.1.1
kombu==4.6.7
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.18.0
more-itertools==7.2.0
multidict==4.5.2
mysqlclient==1.4.6
nodeenv==1.3.3
numpy==1.17.3
ordereddict==1.1
pandas==0.25.3
pendulum==1.4.4
pre-commit==1.18.0
prison==0.1.0
psutil==5.6.5
pycparser==2.19
pyee==6.0.0
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.10.0
pyppeteer==0.0.25
pyrsistent==0.15.5
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
selenium==3.141.0
setproctitle==1.1.10
six==1.13.0
slackclient==1.3.2
soupsieve==1.9.5
SQLAlchemy==1.3.10
tabulate==0.8.5
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.11.0
toml==0.10.0
tqdm==4.40.2
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.6
vine==1.3.0
virtualenv==16.7.9
websocket-client==0.54.0
websockets==8.1
Werkzeug==0.16.0
WTForms==2.2.1
yarl==1.3.0
zipp==0.6.0
zope.deprecation==4.4.0
